# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Profy Skill Academy API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: UNLICENSED
  description: |
    REST API for the Profy Skill Academy learning platform (TECHNICAL_DOC §5).

    * Consumer endpoints (mobile + web share them) accept a JWT with audience
      `consumer`. Admin endpoints live under `/api/v1/admin/*` and accept a JWT
      with audience `admin`. The two token types are never interchangeable.
    * Errors use the envelope `{"error": {"code": string, "message": string}}`
      with stable machine codes (`ai_unavailable`, `rate_limited`, ...).
    * Clients send an `X-App-Version` header; the API can force-update later.
servers:
  - url: http://localhost:8080
    description: Local development (docker compose)
...
paths:
...
  /api/v1/taxonomy/tree:
    get:
      tags: [taxonomy]
      summary: Full Phase-1 tree (Phase-2 nodes are never serialized)
      operationId: taxonomyTree
      security: []
      parameters:
        - name: phase
          in: query
          schema:
            type: integer
            enum: [1]
          description: Reserved; the server always filters to phase=1.
      responses:
        "200":
          description: Recursive tree of Phase-1 nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxonomyNode"
  /api/v1/taxonomy/nodes/{slug}:
    get:
      tags: [taxonomy]
      summary: One node with children and breadcrumb
      operationId: taxonomyNode
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ----------------------------------------------------------------- discovery
...
  /api/v1/admin/taxonomy:
    get:
      tags: [admin]
      summary: List all nodes (Phase-2 included — admin sees everything)
      operationId: adminListTaxonomy
      security:
        - adminAuth: []
      responses:
        "200":
          description: Flat list of all nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxonomyNode"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [admin]
      summary: Create a category/subcategory/course node
      operationId: adminCreateTaxonomyNode
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyNode"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Slug already exists or cycle/depth guard failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/taxonomy/reorder:
    post:
      tags: [admin]
      summary: Reorder siblings (sort_order batch update)
      operationId: adminReorderTaxonomy
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "204":
          description: Reordered
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/admin/taxonomy/{id}:
    get:
      tags: [admin]
      summary: Get one node (with children)
      operationId: adminGetTaxonomyNode
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyNode"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [admin]
      summary: Update a node (name, slug, description, phase, active, sort)
      operationId: adminUpdateTaxonomyNode
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyNode"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Slug conflict or cycle guard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [admin]
      summary: Delete a node (fails if it has children or lessons)
      operationId: adminDeleteTaxonomyNode
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # -------------------------------------------------------------- admin lessons
...
components:
  securitySchemes:
    consumerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Consumer JWT with audience `consumer` (15 min).
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT with audience `admin` (15 min). Never interchangeable with consumer tokens.
  responses:
    Unauthorized:
      description: Missing/invalid/expired token, or token audience mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found (or invisible to this audience)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Invalid request body/parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine code, e.g. ai_unavailable, rate_limited.
            message:
              type: string
...
    TaxonomyNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: [string, "null"]
          format: uuid
        node_type:
          type: string
          enum: [category, subcategory, course]
        name:
          type: string
        slug:
          type: string
        description:
          type: [string, "null"]
        icon:
          type: [string, "null"]
        phase:
          type: integer
          enum: [1, 2]
        is_active:
          type: boolean
        sort_order:
          type: integer
        depth:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/TaxonomyNode"
    NodeDetail:
      allOf:
        - $ref: "#/components/schemas/TaxonomyNode"
        - type: object
          properties:
            breadcrumb:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  slug:
                    type: string
    TaxonomyCreateRequest:
      type: object
      required: [node_type, name, slug]
      properties:
        parent_id:
          type: [string, "null"]
          format: uuid
        node_type:
          type: string
          enum: [category, subcategory, course]
        name:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
        icon:
          type: string
        phase:
          type: integer
          enum: [1, 2]
          default: 1
        is_active:
          type: boolean
          default: true
        sort_order:
          type: integer
          default: 0
    TaxonomyUpdateRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: [string, "null"]
        icon:
          type: [string, "null"]
        phase:
          type: integer
          enum: [1, 2]
        is_active:
          type: boolean
        sort_order:
          type: integer
    ReorderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, sort_order]
            properties:
              id:
                type: string
                format: uuid
              sort_order:
                type: integer

    # -------------------------------------------------------------- content
...